requirements.txt
streamlit
openai
python-dotenv
PyPDF2
langchain
langchain-community
numpy
scikit-learn
tqdm
